requirements.txt
flask==2.3.3
gunicorn==21.2.0
huggingface-hub>=0.19.4
diffusers==0.23.0
transformers==4.35.2
accelerate==0.25.0
safetensors==0.4.1
pillow==10.1.0
torch==2.2.1
torchvision==0.17.1
torchaudio==2.2.1
bitsandbytes==0.41.1
scipy==1.11.3